chore: cleaned up piegraph name
